Simplified gunshot aim direction code to work consistently with AI characters
Since a gun will always be shot by a Character, in the direction said character is aiming, I made a simple property for obtaining it. I changed the GunGeneralStats.Shoot() function to obtain the origin and direction parameters from the 'user' parameter.

I made this change because the shooting controls previously relied on some values that could only be obtained from a WeaponHandler class. Additionally, this new version of GunGeneralStats.Shoot() only requires one parameter.