WIP: Diegetic light source and visibility meter
The DiegeticLightSource class allows code to check how much an entity in the scene is being illuminated by the lights around them.
I also made a VisibilityMeter class, so the player can see how well-lit they currently are.
That being said, none of my AI functions currently use this, but that will change in future.