Player sprint function
This ability replaces the half-baked 'dodge' mechanic, and allows the player to move considerably faster at the cost of stamina.
It's presently a bit too easy for the player to continuously dodge enemy projectiles, so as well as adding this sprint function I lowered the player's base movement speed. My idea is to make it more difficult to dodge projectiles at normal speed, encouraging use of cover, but allowing the player to make quick dashes out in the open.