AITargetManager class, for tracking an AI combatant's target
This class stores what target the AI wants to attack, and uses a FieldOfView class to continuously update if the AI can see it and its last known position. This will be used for more complex AI behaviours, and to cache the output of the more complex FieldOfView detection checks to improve performance.