StateController and StateFunction system, as a simpler alternative to my behaviour tree system
That system proved to be far too difficult to work with, due to the ScriptableObject-based nature making it impossible to establish references to different scene elements. This system I have created just links between different MonoBehaviours and uses Unity's inbuilt execution order, so while I do have to create a lot of child objects to contain scripts, it also makes it much easier for my AI to switch between the current active state.