WIP: refactoring AI states to switch between combat and searching more effectively
I changed the EngageTarget function to inherit from AIStateFunction instead of the AIState class. I also added a reference to an AIGunAttack, so EngageTarget can enable/disable that attack behaviour when necessary.
I've also started work on a 'SearchForTarget' function, where the enemy travels to the player's last-known position and looks around to find them. This isn't complete (I currently have a much simpler version for the sake of testing and ensuring the whole conversion didn't completely break the system), but it's good progress towards far smarter and more intuitive enemy AI.