WIP: 'strawberry jam' screen overlay to indicate low health
This animated overlay is a placeholder and may be changed to a different indicator, but I felt a feature like this was necessary. When playtesting, I found myself playing far too recklessly with low health and repeatedly being killed, because my brain wasn't fully registering that I had very little remaining health.
Additionally, a bug is present that causes the overlay to go away if the player pauses and unpauses the game, which I need to fix.