Replaced 'SearchForTarget' with 'SearchForLastTargetPosition' class
The former was a placeholder, while I worked on the more accurate SweepAreaForTarget class. Now that that works, I have removed it and implemented the better 'SearchForLastTargetPosition' class in its place, where the AI goes to the last-known target position and looks around. However, it is not entirely complete, since the AIAim.SweepSurroundings() function is not completely working yet.