WIP: Smoke clouds!
Time for an actual unique gameplay mechanic!
I created the new 'SmokeCloud' class, that creates physical colliders and matches up their positions and sizes with the particles created by a ParticleSystem. This allows the particles to function like fully-fledged colliders and block raycasts.
After updating the layers and collision detection to ensure that the player and enemy vision detects the smoke, but their weapons don't, I have created a functional plume of smoke that moves and dissipates in an intuitive and semi-realistic manner, while effectively obscuring enemy vision.
I need to improve the visuals at some point, most likely using a fancy shader, but the placeholder appearance is close enough that the player should be able to intuit what it is.
I intend to capitalise on this as a full-fledged gameplay system, with smoke grenades, fires, shootable fire extinguishers, wind, and other ways to use smoke to control visibility.