WeaponHandler - quick switch now changes the firing mode on the current weapon
There was a bug preventing this from working correctly, but it has since been fixed. I have included a bit of code so it can be switched to still change weapons, although I have not implemented a toggle in the player's options menu for this.

This was a feature highly requested by playtesters, and now it's finally working as intended!