WeaponHandler - added support for off-hand attack
I added a new input, a reference to an 'off-hand' attack in the WeaponHandler, and I refactored the 'Weapon' and 'WeaponMode' classes so I can activate them using different inputs, allowing me to re-use the same weapon code for other attacks. Now the player can activate an off-hand ability by pressing the 'G' key, or the left shoulder button on their gamepad.