'ThrowObject' attack type, and smoke grenades!
The 'ThrowObject' class and 'Throwable' projectile type will be used for all weapons that the player will throw. It currently doesn't support animations or delays between throws (so you can mash the button and eject as many as you want at lightning speed).
The first attack that uses it are smoke grenades. The player can chuck out a grenade that uses the new 'SmokeCloud' class to spray a plume of smoke for several seconds, obscuring player and enemy visibility. I want to do some more playtesting to ensure the smoke feels just right, but I'm extremely pleased with how everything looks and feels so far.