Optimised 'look around for last target' code to prevent infinite looping, and general cleanup
The 'QuickLookForLostTarget()' IEnumerator was only being used in the SearchForLastTargetPosition class, and I realised that's the only place where said code would need to be used, so I moved it there instead to make the flow of logic more intuitive.
I also realised the while loops in this code must have been the cause of frequent freezes I noticed in gameplay, and that the code could be changed to remove all of them without sacrificing anything.