SweepAreaForTarget - optimised NavMesh path check
Previously, when a new set of points were obtained to check, the code would calculate a path for all of them and eliminate all points the AI couldn't reach. This caused a severe lag spike whenever an enemy would switch to this state.
I've fixed this by only having the AI check the paths when deciding on a new destination, only removing points that it was planning on going to immediately. This spreads the checks out over multiple frames, and also ensures points are eliminated on their own through the regular search function and don't need to have paths calculated in the first place.