Sprint function cleanup
I changed the 'time standing still before marked as stopped' delay to nothing. This also uncovered a bug preventing the sprint from triggering at all if this was the case. Additionally, I took the RegeneratingResource class used for stamina and placed it on its own GameObject, as before there was no way to tell that it was supposed to be for the player's stamina.