WIP: objective system for player progression
The system is fairly simple - place an ObjectiveHandler in the level and add a bunch of objectives as children, and the game will set them up and record progress. Some objectives are mandatory, some are optional, and some require other objectives to be completed first. All standard stuff.
The three current objective types are for killing a list of enemies, moving to a specific location, or interacting with an object.
This system functions, but is still a work-in-progress. For example, nothing currently happens once all the objectives are completed.