WIP: Diegetic sound system for AI to hear sounds in the level
Enemies will have a 'DiegeticAudioListener' class, that plays an event when a sound occurs close enough for them to hear it. They can then perform different actions based on what kind of sound was played.
I'm using the RandomSoundPlayer class (which I have now renamed to DiegeticSound) to differentiate certain sounds, because I'll be able to drag and drop specific sounds for enemies to respond to, and always ensure that a sound the player hears matches up with what the enemy hears.
This system is still a work-in-progress, it's currently untested and there is no code that makes use of it.