Level completion screen
Once the player has completed all objectives, a screen will display, showing their stats and bonus objectives, and showing buttons to either proceed to the next level or quit to the menu. Fairly simple stuff and I intend to expand on it in future, but it currently works well for providing structure to the game level.