Added extra rooms and enemies to the level, to make it feel more complete
I populated the multi-level 'cell blocks' area with enemies, and added a final hall with several enemies in different places and some vents to hide in. After that, there's an upwards passage, and an end-of-level marker to form a satisfactory conclusion. The scene needs some more tweaks to be fun, but it does feel like a complete level now.