Tweaked offhand attack code so I can assign more than one in a modular fashion
I created an 'OffhandAttack' class that allows me to specify a weapon mode and an input action, and will automatically set up a listener to trigger the attack. This ensures I don't have to hardcode function names for each input value, and that the script doesn't need to be on the same gameobject as the PlayerInput.