Refactored look controller to be less janky and make it easier for external scripts to alter the direction a character is looking in
The player's base movement controls are some of the earliest features I implemented, and therefore some of the jankiest. The look controller in particular used to store some pre-defined angle values and then update the transform orientations accordingly every frame, and so externally updating the transforms would have no effect. In order to make a melee attack script that automatically shifts the player's aim towards a target, I needed to refactor the whole system so that the current look values as used by the player's look controls directly represented the transform values, allowing them to be altered externally without having to perform any unnecessary extra math or variable assignment.