WIP: player melee attack
The player can now press 'F' on the keyboard to perform a kick, playing a small animation and dealing damage to the nearest enemy within a specified attack angle and range.
The system works reasonably well but has a lot of changes and fixes that need to be made. There are no sound or particle effects, and the attack doesn't apply knockback or detect surfaces other than entities, so those need to be fixed.