Melee attack now registers terrain and props if it doesn't specifically hit an entity
I created a 'DamageDealer' class designed to store all the information for damage stats and on-hit effects, that can be run in a self-contained package to ensure the various effects play no matter what the attack hits. I also intend to refactor the Projectile class to use this new system, as that class also uses an overly complicated way to run its various hit effects.
I also gave the melee attack some placeholder noises for swings and impacts , but they're not quite punchy enough and need to be replaced later on.