Stamina cost for melee attacks, and WIP on overhaul to weapon attack requirements
The first bit is self-explanatory so I'll go to the second bit. The code for determining what needs to be valid for an attack to occur, plus what happens after, is extremely messy and designed only for ranged weapons. It's only become a problem now because I'm expanding into other kinds of weapons. I'm making steps towards standardising a lot of the attack code, so functionality that exists within RangedAttack will apply to MeleeAttack and ThrowObject too.