PhysicsAffectedAI script, so AI enemies can now be affected by physics forces without compromising their pathfinding
This class prevents the NavMeshAgent from automatically updating the transform's position. Its own position is instead synced to match the object's transform, and the desired velocity is acted upon by moving the AI using Rigidbody.MovePosition(). This allows the same movement as before, but the AI will now move when acted upon by outside forces. The agent will realistically move to counter the knockback inflicted upon it, allowing it to easily return to normal movement after being shifted by an impact.
I'm surprised at and pleased with how easy this was to implement, and how smoothly it appears to run.