Enemy stun mechanics!
If an enemy has a CharacterPoise script attached, taking damage will fill up an invisible meter. Maxing this out will cause the enemy to be momentarily dazed (by switching to a different state), interrupting whatever they're doing and allowing the player to score some free hits.
This mechanic will act as another way for the player to manage numerous or more powerful enemies, either by stunning them to disable their attacks or selectively disabling certain enemies to focus on the rest of a group.