WIP: Diegetic audio (and new ImpactEffect class)
Enemies can now hear sounds! I added a base ‘decibels’ value to the DiegeticSound script, and I refactored a lot of the audio code so that when a sound is played it signals any DiegeticAudioListeners in the scene, and also to clean it up generally.
I also simplified the detection for now, to be just based on distance (with falloff calculated using the inverse square law).
I also made a new ImpactEffect class. This pairs up a particle system and sound effect, and will eventually be used to play different effects for different surfaces. This system is currently only used for bullet impacts, but will eventually be used for footsteps, physics impacts and various other instances of things hitting terrain or other objects.