EnvironmentalHazard class for objects that instantly kill players and enemies
This class is very simple - if a Character hits the attached collider (or enters it, if it's a trigger) they take a massive amount of damage and die instantly.
This script will be used for environmental features like pits, cliffs, grinders, and anything else that a person wouldn't survive falling into.