Simplified InverseSquareValueMultiplier to be simpler to use, and protected against divide-by-zero events
Instead of directly multiplying a base value, the function returns a value that can be multiplied onto any other number for the same effect. The value is also never zero or less, and it returns 1 as a safeguard if the distance value is zero, so a divide-by-zero error will never occur regardless of the parameter, as the user doesn't need to do any division themselves.