Enemies now investigate suspicious noises (also refactoring of AITargetManager and SearchForLastTargetPosition)
If an enemy hears a noise (and aren’t busy engaging the target or searching for something more important), they will go to that position and look around, returning to their previous routine if they don’t find anything interesting.
To make this change I took most of the code from ‘SearchForLastTargetPosition’ and moved it into a more general ‘InvestigateLocations’ class, so the AI can be set to search for any kind of suspicious noise. This system also uses a priority value to ensure the AI doesn’t rapidly switch their search target, instead only doing it if a new target has a higher priority. For sounds, priorities are handled by the order of sounds in the new new ‘InvestigateSuspiciousNoise’ class, which listens for an ‘onSoundHeard’ event and sends a search request to InvestigateLocations.
I also refactored ‘AITargetManager’ to directly include code for switching to the attack state if it finds a target, meaning that I don’t have to include it in every function where the AI could possibly be interested in finding a target. This also happened to allow me to clean up ‘SweepAreaForTarget’ by a small amount.