Refactored gun reticle to more accurately represent the width of a standard target, at the maximum distance where the current weapon's shots won't miss.
Now the player can accurately aim their reticle over a standard-sized enemy, and if the reticle is the same size as their hitbox or smaller it means the shot is guaranteed to hit.
I also changed the reticle shape to something more standard and less overdesigned, for even more visual clarity.