WIP: Test AI now has an actual ragdoll and animations
There are still bugs to fix and tweaks to make, but it's still a significant milestone.
One change I had to make was adding a separate 'physics collider' to the enemy, on a layer that's ignored by attacks but is still registered for physics collisions. This collider is automatically updated by PhysicsAffectedAI, to conform to the NavMeshAgent size and be enabled/disabled when needed.