Fixed bug where AI field of view would be blocked by their own bodies
I tweaked the vision cone code to include extra checks, to ignore the AI's own colliders. This did require me to make a slightly different version of the existing line of sight code, designed to check for a specific target and not just if an unobstructed line can be drawn between two points.