Enemies now hold their guns properly (and drop them on death)
Enemies holding their guns properly was as simple as changing the weapon's parent to the enemy's hand. Having it drop on death was more complicated, and I relied on a somewhat janky solution. I created a 'dropped rifle' item that functions like an ammo pickup, which spawns in the exact same position as the enemy's rifle while the original is hidden. I made this workaround as the weapon objects are somewhat messily set up, and originally designed only to be used by the player. This current system may become redundant in future, as I might refactor the way the weapon transforms and animations are set up.