Added impact noises for some props (including ragdolls)
I created the 'PhysicsObjectEffects' class which stores impact info on an object that might have colliders attached as children. When an impact occurs, it plays an ImpactEffect at the points of contact, at an intensity proportional to how hard the object was hit. This allows physical objects to feel much more 'physical' by visually and audibly responding to impacts.