ImpactEffect now natively supports applying decals
Instead of needing to attach a decal effect onto the particle effect, I modified ImpactEffect so a Sprite can simply be assigned in a new decal field. A decal renderer prefab is then spawned using ObjectPool, given the correct sprite, and positioned in the correct spot. This means decals don't have to be awkwardly added onto the assigned particle effect.