PhysicsAffectedAI - ragdolls now properly preserve initial momentum when ragdollised by an attack
Initially, the process of applying physics on a killing blow, dealing the actual damage, triggering the death event and activating the ragdoll would all happen within a single frame. This means that the applied force from the killing blow would fail to transfer to the rigidbody. This is because when applied with any variation of AddForce, the code has to wait for the next FixedUpdate call for that change to be fully processed and shown in the velocity values. To fix this, I changed the 'change ragdoll active status' code to an IEnumerator, that yields until the next FixedUpdate call before activating/deactivating the ragdoll.
Also, existing ragdoll momentum is now transferred back to the base rigidbody when reverting from ragdoll to regular physics.