New placeholder shotgun enemy
Compared to the rifle enemy, this new enemy moves slightly faster, pursues the player more aggressively and attacks with a shotgun (that when killed, can be picked up for ammo). Overall health and hitbox sizes remain the same, however.
The rifle enemy's AI has also been tweaked slightly to make them more cautious and keep a longer distance, to provide a better contrast.