New 'SingleInput' class to more effectively assign input listeners to specific objects
This will allow me to put all kinds of input-accepting functions onto the player, without them having to be on the same GameObject as the PlayerInput component.
I re-used the code from OffhandAttack and made it more universal, then refactored that class to use this new one, preventing boilerplate repeating of the same functionality.