Refactoring of interaction function - GUIs are improved and the player no longer needs to directly look at an object to interact with it
This is a little QOL effect so the player can interact with things more quickly in a fight, e.g. to pick up more ammo. That sort of thing shouldn't actually be difficult to do, so having it be finicky isn't 'challenge', it's just annoying.
The new interaction system is also designed to work with my upcoming mechanic for picking up and throwing objects.