Entities won't take physics damage when hit by objects they just launched themselves
I added a second dictionary check similar to the 'recentPhysicsCollisions' check, that's instead updated by DamageDealer or ThrowHandler whenever an entity uses one of those to apply force to a physics object.

This ensures that if the player throws a ragdoll and it hits them due to glitching out, the player is less likely to take damage. They'll still take damage after the cooldown though, since there are still lots of situations where they would realistically be hurt by an object they threw, e.g. flying into the air and landing on their heads.