Improved ragdoll joints
Replaced all joints with ConfigurableJoints, for ease of adjusting settings.
Removed joints and rigidbodies for neck and shoulders. The shoulder joints were always fixed in place and had no reason to have their own physics, and the neck having a separate joint from the head resulted in the head having too wide a range of movement.
Enabled projection and disabled preprocessing, to limit impossible stretching.