Refactored StateController to have hierarchical functionality.
StateControllers now act as their own states and can be switched to from parent StateControllers, resulting in an easy hierarchy system. If a state is switched to, all parent states will switch accordingly to ensure that the desired state is accessed. This allows easy compartmentalisation.
The new StateController now also knows its immediate children, and the root of the hierarchy.
I also added a UnityEvent that triggers when a StateController is enabled or disabled, so other scripts can be enabled or disabled along with it.