Fixed bug where ragdollised enemies couldn’t be killed by environmental hazards, due to the root entity not being found
To fix this, I took several disconnected functions for ‘getting an entity/character reference from a child object’ and combined them together in EntityCache. Now that class’s function will check if the child gameobject is part of a Ragdoll. If so, it’ll get the ragdoll’s entity reference instead.
Additionally, I changed EntityCache's dictionary to only require a GameObject as the key, instead of a Collider, since the value is only used as a unique reference and doesn't rely on any of the Collider class's unique features. This let me get rid of some separate functions that existed just to get an Entity reference from a rigidbody instead of a collider.
I also changed several parts of the project to reference the base Entity type instead of the more specific Character type, as that specificity wasn’t needed (except for one check to ensure the player can’t pick up an alive character).