WIP: shootable fire extinguishers
I intend for the player to be able to shoot these to rupture the casing, releasing blinding clouds of powder (like a classic exploding barrel but for smoke), as well as pick up and throw them for clever positioning.
I have to fix the particle effect to play properly, as well as ensuring it shoots out in a way that's both realistic and fun for gameplay. I also need to add damage resistance code to ensure it can only be damaged by bullets, so it won't explode just from being thrown for repositioning.