Fixed bugs where enemy AI would reanimate after death or stunning
This was due to the AITargetManager and DiegeticAudioListener continuing to run, and triggering state changes that should only be allowed while the character is actually alive. I added listeners to change their active states based on the 'active' state enabling and disabling.
This in turn required me to change some component references - the base state machine had as its 'current' state one that wasn't an immediate child, which caused some errors, so as well as manually changing the values, I added some code to automatically change invalid ones.

With all that, the enemy AI now stays dead, and only recovers from stunning when the stun state's own code says so.