Ragdoll animator reference cleanup
The ragdoll has no reason to reference an animator, except to enable/disable it. I realised that code could be added from HumanoidAnimator as a listener. This makes it easier to control those variables, and makes the ragdoll class simpler.
I also removed an unnecessary reference in HumanoidAnimator to said ragdoll, and instead made the variable into a property referencing the value in PhysicsAffectedAI.