Refactored hitbox health assignment, to prevent the player from taking damage when throwing corpses
Sometimes, hitboxes on thrown corpses (which have Hitbox scripts but no Health script) would get confused and assign themselves the player's health script, when the player picked them up.
I fixed this by changing how Hitbox's 'sourceHealth' and 'attachedTo' properties work. Instead of searching for a health script, and then referencing the entity the health script is attached to, it instead checks for an entity, and returns the entity's health script. This is because all objects with health meters have Entity components, but not all entities have health values.