Enemies now smoothly interpolate from ragdoll state to standing up animation
This new feature required a bunch of changes and bugfixes, spanning over several scripts for both the AI and animations. The code involves storing the  But it helps the enemies look somewhat less janky. There’s still a bit of jank, but for a game that emphasises physics interactions, developed by one guy, I’m willing to tolerate a bit of jank. And the jank is often part of the fun!