Fixed bug causing overly jittery enemy ragdolls in build
I realised the mass averaging code was running twice (once in PhysicsAffectedAI and another time afterwards in Ragdoll itself), and the second time it was using one of its own rigidbodies' mass as reference. This resulted in the ragdoll thinking its mass needed to be far lower than it was, and making itself too light and jittery.
I then added a new component to update the combined mass of solitary ragdolls, that's separate from the Ragdoll script itself (so there's no code for it in Ragdoll to interfere with PhysicsAffectedAI).
While I was at it, I also changed my ragdoll's layer to be 'Entity' instead of 'Default', ensuring I didn't need to override it in the enemy prefab.