Simplified rifle scope to sidestep several bugs
I wanted to be fancy and have a second rendered scope view while still showing the surrounding environment at standard zoom, but there were several bugs:
•	The scope would sometimes clip through walls due to how the secondary camera was positioned in world space.
•	The render texture outright fails when I run it on my Steam Deck (not sure if this is because of Linux or the Steam Deck specifically).
I couldn’t be bothered debugging the janky shader I made for it, so I opted for the standard overlay effect seen in many other first-person shooters. As well as just zooming in the main camera, an overlay appears to show the scoped-in view, and the weapon model will become invisible. A quick black flash exists to mask this transition.
This method is less impressive to brag about, but it still looks great while being far more functional and optimised.