Hitboxes can now be more resistant or vulnerable to different damage types
By creating a DamageResistanceProfile and applying it to the Hitbox, the damage code can now check if a particular kind of attack should deal more or less damage. If one isn't applied, the code functions the same as before.
I accidentally added an unfinished version of DamageResistanceProfile in a previous commit, but here's the functional version of it.