Shootable steam valves (and DamageOnInteract) script
These valves can be damaged (or just interacted with) to release a burst of steam, similar to the shootable fire extinguishers.
I created a new script that lets me set up context-sensitive interactions to damage things. This allows the player to quietly activate the steam valve by turning it, instead of attacking it.