Breakable wooden wall (and gibbable objects)
This simple wall can be broken down with attacks to continue. It's basically to tutorialise kicking and throwing objects.
To accomplish this I made a new 'GibbableObject' class, that handles unparenting the gibs and activating their physics. I'll have to expand on this class in future, so the gibs respond better to physics forces when initially broken.