WIP: new tutorial level
This new demo level exists to teach the player each mechanic in order, and better show them how the game is meant to be played. The player starts with no weapons and must make extensive use of thrown objects, environmental hazards and smoke to take out their enemies.
The level is nowhere near complete, as I'm still figuring out the structure, but it's turning out well so far.