Enemies can now be stunlocked
The stun meter is reset when the enemy is knocked down. When the enemy takes enough stun damage to bring the meter back up to a stunlock threshold, the knockdown procedure is reset. This will allow the player to easily take out certain isolated enemies, and also create more funny ragdolls.