Fixed bug preventing AI hitboxes from being registered
Turns out the AI death state was triggering once upon spawning, even if the AI was still alive. This detached the ragdoll and prevented the Health script from finding said hitboxes.
I fixed this with a simple check to not run the death state code if the AI isn't actually dead. In the process of debugging, I also cleaned up the hitbox and collider obtaining code to be cleaner to read.