FieldOfView - changed code to still return hit data even on a failed check
This ensures the AI can still act on whatever info they do have (which is good for debugging).
I also updated the target manager code to save that info separately. Since its code relies on saving the last time the player was seen, we don't want that overridden by data on a different thing blocking vision. So I renamed 'lastHit' to 'lastValidHit' and added a new 'lastHit' that records whatever data is seen regardless of if it's actually the desired target.