StateController.root now returns an actual StateController
It being the root implies it controls states, and would only be referenced in order to gain that controller data. It wouldn't even be helpful to return a state that isn't a StateController, in the rare event of a state that runs without being connected to one. So I tweaked the property to return null if the current state has no parent controller, and isn't a controller itself.