Fixed bugs where enemies can revert to being idle, or get stuck in inactive limbo states, partway through investigating suspicious occurrences
The reverting to idle was because of ‘onFail’ being assigned right before switching to said state, and it also being assignable from InvestigateSuspiciousNoises. This meant that before certain functions could trigger the switch, InvestigateSuspiciousNoises might trigger a check of its own with its own fail state. This would override the intended check fail state.
The ‘stuck in limbo’ effect was related to SearchForLastTargetPosition. Said code involved triggering an InvestigateLocations check, which then got shut down because the cooldown hadn’t elapsed since the last check. That cooldown is meant for short-term occurrences like suspicious noises, but not for important procedures that act as an essential part of another logic state. I fixed this by adding an option to override the cooldown.