Reduced player sprint speed, to make it harder to dodge attacks
This seems like it might go against the recent changes to the enemy AI, to make their attacks slower and more telegraphed, but I want to change the nature of the fights. The combat previously relied more on twitchy reactions, and was more fast-paced akin to a regular shooter. Since the player can't dodge enemy attacks as easily, but they take longer to execute and have more obvious telegraphs, it slows the gameplay down resulting in a more tactical feel.