WIP on new tutorial level combat encounters, and tweaking of level components
I tweaked the dimensions of the stairwell assets to make them easier to design layouts around, and added a third place to add a doorway. (I then updated the original level to accomodate this tweak, as some rooms needed to be shifted.
I also created a 'fire extinguisher mount' asset that can be easily placed on walls at an appropriate human height, making it easy to place fire extinguishers around levels in a diegetic manner.