Fixed bug where interaction handler would sometimes ignore objects the player is directly looking at
I fixed this by adding a preliminary raycast straight forward, to check for an interactable object. If this doesn't return anything, then it goes to the 'scan objects within a distance and check for line of sight' code.
This ensures the player can always reliably select an object by directly aiming at it, regardless of what other stuff is positioned around it.