Weapon pickups can now be ID-locked, preventing player pickup
This allows me to prevent the player from getting a weapon until the point that I want them to, while still allowing enemies to have said weapon. Once the player has said weapon, these drops will function like normal ammo pickups.
I may not commit to this method in the final game, since I'll have more enemies for the player to fight that have different weapons. But it's a good solution for now.