Doors
I've finally created doors! They can be opened or closed by interacting with them, and also behave as physics objects on their hinges.
Doors can also be locked to prevent progression, and unlocked using a DoorLock object (this is an abstract class, so I can add new kinds of locks in future).