Fixed bug where enemies couldn’t take fall damage while animating
This was due to the hitboxes being kinematic for the purposes of animation, and always showing a velocity of zero even when the parent rigidbody is moving at an extremely high speed. When hitting level geometry (which doesn’t move at all), there’s no velocity to record.
If the recorded velocity has a magnitude of near zero, a function now runs that checks in both colliders for non-kinematic parent rigidbodies, and calculates the relative velocity off those values instead. If only one rigidbody is present, it uses just that velocity value instead (and mirrors if necessary to still accurately represent relative velocity). A dot product check is also triggered against the collision normals, to ensure the two objects actually collided and aren’t somehow moving away from each other.