Replaced handgun's full-auto mode with a 3-round burst
The full-auto mode was overpowered in the right hands, and extremely janky in the wrong ones. This 3-round burst serves a more distinct and balanced purpose. Being able to fire 3 shots extremely quickly allows for stunning enemies with a single burst, or one-shotting enemies with a burst to the head, but the high ammo consumption and cooldown means the mode requires more precision aim.
To highlight this difference I also reduced the semi-auto mode's base fire rate - not enough to feel any different, but enough that players can't fake full-auto using some kind of auto-fire program.