WIP: simple checkpoint system
If a player dies on a level, they'll now be teleported to the furthest checkpoint they touched and get some of their health back. A death counter will also be incremented.
A full-on saving system will be a time-consuming endeavour for a later date, but sending the player all the way back to the start of the level was too punishing. So this method is a simple stop-gap. The death counter ensures that dying still has some kind of consequence, as players can challenge themselves to complete the level without dying.