New objective to complete a level without dying
This works with the CheckpointManager to show how many times a player has died throughout the level, and reward them if they avoid doing so.
This objective will be removed once I add an actual save/load system, as later on the player will actually die and have to start from their last save.