New title and logo
'Smoke 'Em If You Got 'Em!'
I'm quite pleased with the logo. It'll likely be changed in the future, but it's better looking than what came before, and is an important step in helping this game form an actual identity. I changed the main menu and metadata to also match these changes.