Updated InteractionHandler to use the same universal 'check within angle' code used for the player's aim assist.
The general functionality is the same (aside from some minor differences that don't really matter in the grand scheme of things), but since similar code is now being used in multiple parts of the app, I want to reduce the amount of boilerplate. I also plan to update the melee attack hit detection, to use the same detection code as well.