Converted project to use URP rendering, and WIP on toon shader for a more unique visual style
As well as learning shader skills to improve my resume, I'm starting on the steps towards the desired visual style, and implementing several mechanics that are heavily reliant on shaders. The main change for now is the new toon/outline shader I'm using, to render everything in a comic book-esque style. I will have to tweak this shader until (A) I can get the outlines right, and (B) I can ensure the lighting visually matches the light detection I intend to have for enemy AI, ensuring the player can clearly tell if they're visible or not.